+++ internal/kms/root_key_test.go
@@ -30,7 +30,7 @@ func TestRootKey_Create(t *testing.T) {
 		args          args
 		want          *kms.RootKey
 		wantErr       bool
-		wantIsErr     error
+		wantIsErr     errors.Code
 		create        bool
 		wantCreateErr bool
 	}{
@@ -38,7 +38,7 @@ func TestRootKey_Create(t *testing.T) {
 			name:      "empty-scopeId",
 			args:      args{},
 			wantErr:   true,
-			wantIsErr: errors.ErrInvalidParameter,
+			wantIsErr: errors.InvalidParameter,
 		},
 		{
 			name: "valid-org-config",
@@ -72,7 +72,7 @@ func TestRootKey_Create(t *testing.T) {
 			got, err := kms.NewRootKey(tt.args.scopeId, tt.args.opt...)
 			if tt.wantErr {
 				require.Error(err)
-				assert.True(errors.Is(err, tt.wantIsErr))
+				assert.True(errors.Match(errors.T(tt.wantIsErr), err))
 				return
 			}
 			require.NoError(err)